third_party/WebKit/LayoutTests/fast/canvas/canvas-createImageBitmap-invalid-args.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<script>
description("Ensure correct behavior of createImageBitmap for invalid inputs.");
window.jsTestIsAsync = true;

var InvalidStateError = "InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable.";
var TypeError = "TypeError: Type error";
var IndexSizeError = "IndexSizeError: Index or size was negative, or greater than the allowed value.";

function shouldNotBeCalled(imageBitmap) {
    testFailed("shouldNotBeCalled was called.");
}

var image;
var testBitmap; // this is an ImageBitmap that is uncropped. We use this to test createImageBitmap(testBitmap)
var d;          // image.imageData

// Create auxiliary canvas to draw to and create an image from.
var aCanvas = document.createElement("canvas");
aCanvas.setAttribute("width", "200");
aCanvas.setAttribute("height", "200");
var aCtx = aCanvas.getContext("2d");

image = new Image();
image.onload = imageLoaded;

// Before image loads
shouldThrow("createImageBitmap(image)", "InvalidStateError");
image.src = aCanvas.toDataURL(); // set a data URI of the base64 enconded image as the source

video = document.createElement("video");
video.addEventListener("canplaythrough", videoLoaded, false);

// Before video loads
shouldThrow("createImageBitmap(video)", "InvalidStateError");
video.src = "../../compositing/resources/video.ogv";

var imageLoaded = false;
var videoLoaded = false;
var imageBitmapLoaded = false;

function imageLoaded() {
    createImageBitmap(image).then(imageBitmapLoadedCallback, function() {
        testFailed("Promise was rejected.");
        finishJSTest();
    });
    d = aCtx.getImageData(0, 0, 200, 200);
    imageLoaded = true;
    loaded();
}

function videoLoaded() {
    videoLoaded = true;
    loaded();
}

function imageBitmapLoadedCallback(imageBitmap) {
    testBitmap = imageBitmap;
    imageBitmapLoaded = true;
    loaded();
}

function loaded() {
    if (imageLoaded && videoLoaded && imageBitmapLoaded) {
        shouldThrow("createImageBitmap(undefined, shouldNotBeCalled)", "TypeError");

        shouldThrow("createImageBitmap(image, 0, 0, 10, 0)", "IndexSizeError");
        shouldThrow("createImageBitmap(image, 0, 0, 0, 10)", "IndexSizeError");

        shouldThrow("createImageBitmap(video, 0, 0, 10, 0)", "IndexSizeError");
        shouldThrow("createImageBitmap(video, 0, 0, 0, 10)", "IndexSizeError");

        shouldThrow("createImageBitmap(aCanvas, 0, 0, 10, 0)", "IndexSizeError");
        shouldThrow("createImageBitmap(aCanvas, 0, 0, 0, 10)", "IndexSizeError");

        shouldThrow("createImageBitmap(d, 0, 0, 10, 0)", "IndexSizeError");
        shouldThrow("createImageBitmap(d, 0, 0, 0, 10)", "IndexSizeError");

        shouldThrow("createImageBitmap(aCtx, 0, 0, 10, 0)", "IndexSizeError");
        shouldThrow("createImageBitmap(aCtx, 0, 0, 0, 10)", "IndexSizeError");

        shouldThrow("createImageBitmap(testBitmap, 0, 0, 10, 0)", "IndexSizeError");
        shouldThrow("createImageBitmap(testBitmap, 0, 0, 0, 10)", "IndexSizeError");

        finishJSTest();
    }
}

</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
